==> 2_CMIP6Cleaning.Rmd <==
---
title: "2 CMIP6 Cleaning"
output: html_document
date: "2025-08-08"
---

###Manuscript under review

**Title:** Mixed landscapes, social vulnerability, and climate change shape West Nile virus risk across the United States

**corresponding author:** Samantha Sambado (ssambado at stanford.edu)

##Purpose
Process CMIP6 data and format into a shape that works well for R0(T) models. I am working with data for three time periods under two emission scenarios. Time periods are historic/current (2015-2020), mid-century (2045-2074), and late-century (2075-2100). Emission scenarios are SSP2-4.5 (low-emissions) and SSP5-8.5.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

## upload necessary packages
library(raster)
library(sf)
library(exactextractr)
library(tidyverse)

## Spatial boundaries
# US county boundaries
counties_sf <- tigris::counties(year = 2020, cb = TRUE, class = "sf")


# US state boundaries
all_states_sf <- tigris::states(year = 2020, cb = TRUE, class = "sf") # for figures
states_sf <- all_states_sf[!all_states_sf$STUSPS %in% c("PR", "VI", "GU", "MP", "AS", "AK" ,"HI"), ] # omit territories

```

#Step 1. Write function to call in data & process it
```{r}

# Generic function to process one raster brick file (min or max)
process_temp_raster <- function(nc_file, counties_sf, temp_type = c("max", "min"), rcp, century) {
  temp_type <- match.arg(temp_type)
  
  #load RasterBrick
  brick <- raster::brick(nc_file)
  
  #assign CRS and fix longitude extent
  crs(brick) <- CRS("+proj=longlat +datum=WGS84 +no_defs")
  extent(brick) <- extent(brick) + c(-360, -360, 0, 0)
  
  # transform counties shapefile to raster CRS
  counties_transformed <- st_transform(counties_sf, crs = st_crs(brick))
  
  terra::gdalCache(1000)  # Set GDAL cache (memory) to 1000 MB
  
  # extract mean values for each polygon & each raster layer
  extracted_means <- exact_extract(brick, counties_transformed, 'mean')
  
  # add GEOID column
  extracted_means$GEOID <- counties_transformed$GEOID
  
  # convert wide to long format
  long_df <- extracted_means %>%
    pivot_longer(cols = starts_with("mean.X"),
                 names_to = "date",
                 values_to = paste0(temp_type, "_temp")) %>%
    mutate(date = gsub("mean.X", "", date),
      date = as.Date(date, format = "%Y.%m.%d"),
      rcp = rcp,
      century = century,
      # Convert Kelvin to Celsius
      !!paste0(temp_type, "_temp") := .data[[paste0(temp_type, "_temp")]] - 273.15) %>%
    dplyr::select(GEOID, date, rcp, century, all_of(paste0(temp_type, "_temp")))
  
  return(long_df)
}


```

##Step 2. Apply function to each SSP-time period

###SSP245

####Current 
(also called historic, but technically current)
```{r}
# process max temp for 2015-2044 SSP245 historic
max_245_hist <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmax.GFDL-CM4.ssp585.r1i1p1f1.2015-2044.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "max",
  rcp = 245,
  century = "historic")

# process min temp for same scenario
min_245_hist <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmin.GFDL-CM4.ssp245.r1i1p1f1.2015-2044.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "min",
  rcp = 245,
  century = "historic")

# join min & max dataframes by GEOID, date, rcp, century
combined_245_hist <- min_245_hist %>%
  inner_join(max_245_hist, by = c("GEOID", "date", "rcp", "century")) %>%
  filter(date <= as.Date("2020-12-31")) 

#write.csv(combined_245_hist, file = "data/processed/cmip6_ssp245_historic_county_long.csv", row.names = FALSE)
          
```

####Mid
2045-2074
```{r}
# process max temp for 2015-2044 SSP245 mid-century
max_245_mid <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmax.GFDL-CM4.ssp245.r1i1p1f1.2045-2074.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "max",
  rcp = 245,
  century = "mid")

# process min temp for same scenario
min_245_mid <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmin.GFDL-CM4.ssp245.r1i1p1f1.2045-2074.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "min",
  rcp = 245,
  century = "mid")

# join min and max dataframes by GEOID, date, rcp, century
combined_245_mid <- min_245_mid %>%
  inner_join(max_245_mid, by = c("GEOID", "date", "rcp", "century"))


#write.csv(combined_245_mid, file = "data/processed/cmip6_ssp245_mid_county_long.csv", row.names = FALSE)

```


####Late
2075-2100
```{r}
# process max temp for 2015-2044 SSP245 late-century
max_245_late <- process_temp_raster(
  nc_file = "data/raw/cmip6/gfdl_cm4/tasmax.GFDL-CM4.ssp245.r1i1p1f1.2075-2100.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "max",
  rcp = 245,
  century = "late")

# process min temp for same scenario
min_245_late <- process_temp_raster(
  nc_file = "data/raw/cmip6/gfdl_cm4/tasmin.GFDL-CM4.ssp245.r1i1p1f1.2075-2100.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "min",
  rcp = 245,
  century = "late")

# join min & max dataframes by GEOID, date, rcp, century
combined_245_late <- min_245_late %>%
  inner_join(max_245_late, by = c("GEOID", "date", "rcp", "century"))


#write.csv(combined_245_late, file = "data/processed/cmip6_ssp245_late_county_long.csv", row.names = FALSE)

```






###SSP585

####Current 
(also called historic, but technically current)
```{r}
# Process max temp for 2015-2044 SSP585 historic
max_585_hist <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmax.GFDL-CM4.ssp585.r1i1p1f1.2015-2044.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "max",
  rcp = 585,
  century = "historic")

# process min temp for same scenario
min_585_hist <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmin.GFDL-CM4.ssp585.r1i1p1f1.2015-2044.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "min",
  rcp = 585,
  century = "historic")

# join min & max dataframes by GEOID, date, rcp, century
combined_585_hist <- min_585_hist %>%
  inner_join(max_585_hist, by = c("GEOID", "date", "rcp", "century")) %>%
  filter(date <= as.Date("2020-12-31")) 

#write.csv(combined_585_hist, file = "data/processed/cmip6_ssp585_historic_county_long.csv", row.names = FALSE)
          
```

####Mid
2045-2074
```{r}
# process max temp for 2015-2044 SSP585 mid-century
max_585_mid <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmax.GFDL-CM4.ssp585.r1i1p1f1.2045-2074.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "max",
  rcp = 585,
  century = "mid")

# Process min temp for same scenario
min_585_mid <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmin.GFDL-CM4.ssp585.r1i1p1f1.2045-2074.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "min",
  rcp = 585,
  century = "mid")

# join min and max dataframes by GEOID, date, rcp, century
combined_585_mid <- min_585_mid %>%
  inner_join(max_585_mid, by = c("GEOID", "date", "rcp", "century"))


#write.csv(combined_585_mid, file = "data/processed/cmip6_ssp585_mid_county_long.csv", row.names = FALSE)

```


####Late
2075-2100
```{r}
# process max temp for 2015-2044 SSP585 late-century
max_585_late <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmax.GFDL-CM4.ssp585.r1i1p1f1.2075-2100.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "max",
  rcp = 585,
  century = "late")

# process min temp for same scenario
min_585_late <- process_temp_raster(nc_file = "data/raw/cmip6/gfdl_cm4/tasmin.GFDL-CM4.ssp585.r1i1p1f1.2075-2100.LOCA_16thdeg_v20220413.monthly.nc",
  counties_sf = counties_sf,
  temp_type = "min",
  rcp = 585,
  century = "late")

# join min & max dataframes by GEOID, date, rcp, century
combined_585_late <- min_585_late %>%
  inner_join(max_585_late, by = c("GEOID", "date", "rcp", "century"))

#write.csv(combined_585_late, file = "data/processed/cmip6_ssp585_late_county_long.csv", row.names = FALSE)

```
