==> 5_EnvironmentalRiskMaps.Rmd <==
---
title: "5_ProjectedEnviromentalRiskMaps"
output: html_document
date: "2025-08-09"
---

###Manuscript under review

**Title:** Mixed landscapes, social vulnerability, and climate change shape West Nile virus risk across the United States

**corresponding author:** Samantha Sambado (ssambado at stanford.edu)

##Purpose
Create maps of scaled environmental risk of WNV in current, mid- and late-century projections

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)


## upload necessary packages
library(ncdf4)
## upload necessary data
gorris_pip <- terra::rast("data/raw/gorris_sdm/culex_pipiens_meansuitability.nc")
gorris_tar <- terra::rast("data/raw/gorris_sdm/culex_tarsalis_meansuitability.nc")
gorris_quin <- terra::rast("data/raw/gorris_sdm/culex_quinquefasciatus_meansuitability.nc")

cmip6_rR0_allspecies <- read.csv("data/processed/cmip6_rR0_allspecies.csv") %>% dplyr::select(-X)  %>% 
  mutate(GEOID = str_pad(GEOID, width = 5, pad = 0), # make sure FIPS are 5 characters
         GEOID = as.character(GEOID)) 

```


##Step 1. Format data
```{r}
#### Historic / Current
## A. select for only ssp245 historic
rR0_allspecies_245_hist <- cmip6_rR0_allspecies %>% 
  filter(century == "historic", scenario == "ssp245") %>% 
  dplyr::select(GEOID:peakmonths_avgrR0)

## B. make spatial
rR0_allspecies_245_hist_spatial <- counties_sf %>% 
  left_join(rR0_allspecies_245_hist, by = "GEOID") 

## C. break into each species and turn into SpatVec
rR0_pip_245_hist_spatial <- rR0_allspecies_245_hist_spatial %>% 
  filter(species == "Cx. pipiens") %>% 
  vect()

rR0_tar_245_hist_spatial <- rR0_allspecies_245_hist_spatial %>% 
  filter(species == "Cx. tarsalis") %>% 
  vect()

rR0_quin_245_hist_spatial <- rR0_allspecies_245_hist_spatial %>% 
  filter(species == "Cx. quinquefasciatus") %>% 
  vect()

#### Mid Century
## A. select for only ssp245 mid century
rR0_allspecies_245_mid <- cmip6_rR0_allspecies %>% 
  filter(century == "mid", scenario == "ssp245") %>% 
  dplyr::select(GEOID:peakmonths_avgrR0)

## B. make spatial
rR0_allspecies_245_mid_spatial <- counties_sf %>% 
  left_join(rR0_allspecies_245_mid, by = "GEOID") 

## C. break into each species and turn into SpatVec
rR0_pip_245_mid_spatial <- rR0_allspecies_245_mid_spatial %>% 
  filter(species == "Cx. pipiens") %>% 
  vect()

rR0_tar_245_mid_spatial <- rR0_allspecies_245_mid_spatial %>% 
  filter(species == "Cx. tarsalis") %>% 
  vect()

rR0_quin_245_mid_spatial <- rR0_allspecies_245_mid_spatial %>% 
  filter(species == "Cx. quinquefasciatus") %>% 
  vect()



#### Late Century
## A. select for only ssp245 late century
rR0_allspecies_245_late <- cmip6_rR0_allspecies %>% 
  filter(century == "late", scenario == "ssp245") %>% 
  dplyr::select(GEOID:peakmonths_avgrR0)

## B. make spatial
rR0_allspecies_245_late_spatial <- counties_sf %>% 
  left_join(rR0_allspecies_245_late, by = "GEOID") 

## C. break into each species and turn into SpatVec
rR0_pip_245_late_spatial <- rR0_allspecies_245_late_spatial %>% 
  filter(species == "Cx. pipiens") %>% 
  vect()

rR0_tar_245_late_spatial <- rR0_allspecies_245_late_spatial %>% 
  filter(species == "Cx. tarsalis") %>% 
  vect()

rR0_quin_245_late_spatial <- rR0_allspecies_245_late_spatial %>% 
  filter(species == "Cx. quinquefasciatus") %>% 
  vect()
```


##Step 2. Run for each species
Historic/Current
```{r}
## project
risk_pip <- terra::project(rR0_pip_245_hist_spatial, gorris_pip)
risk_tar <- terra::project(rR0_pip_245_hist_spatial, gorris_tar)
risk_quin <- terra::project(rR0_pip_245_hist_spatial, gorris_quin)

## rasterize
risk_pip_raster <- terra::rasterize(risk_pip, gorris_pip, field = "peakmonths_avgrR0", touches = FALSE)
risk_tar_raster <- terra::rasterize(risk_tar, gorris_tar, field = "peakmonths_avgrR0", touches = FALSE)
risk_quin_raster <- terra::rasterize(risk_quin, gorris_quin, field = "peakmonths_avgrR0", touches = FALSE)

## multiple to create scaled risk
scaled_risk_pip <- gorris_pip *risk_pip_raster
scaled_risk_tar <- gorris_tar *risk_tar_raster
scaled_risk_quin <- gorris_quin *risk_quin_raster

## correct CRS
scaled_risk_pip_correct <- terra::project(scaled_risk_pip, "EPSG:4326")
scaled_risk_tar_correct <- terra::project(scaled_risk_tar, "EPSG:4326")
scaled_risk_quin_correct <- terra::project(scaled_risk_quin, "EPSG:4326")

# not sure if i needed to do that
plot(scaled_risk_quin_correct)
plot(scaled_risk_quin)

### make dataframes
scaled_risk_pip_df <- as.data.frame(scaled_risk_pip_correct, xy = TRUE, na.rm = TRUE)
names(scaled_risk_pip_df)[3] <- "scaled_risk"

scaled_risk_tar_df <- as.data.frame(scaled_risk_tar_correct, xy = TRUE, na.rm = TRUE)
names(scaled_risk_tar_df)[3] <- "scaled_risk"

scaled_risk_quin_df <- as.data.frame(scaled_risk_quin_correct, xy = TRUE, na.rm = TRUE)
names(scaled_risk_quin_df)[3] <- "scaled_risk"

## plot them!
ggplot(scaled_risk_pip_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_pip_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 

ggplot(scaled_risk_tar_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_tar_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 


ggplot(scaled_risk_quin_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_quin_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 

## save them
write.csv(scaled_risk_pip_df, file = "data/processed/temp/map_scaledrisk_245_hist_pip.csv")
write.csv(scaled_risk_tar_df, file = "data/processed/temp/map_scaledrisk_245_hist_tar.csv")
write.csv(scaled_risk_quin_df, file = "data/processed/temp/map_scaledrisk_245_hist_quin.csv")
```


Mid century

```{r}
## project
mid_risk_pip <- terra::project(rR0_pip_245_mid_spatial, gorris_pip)
mid_risk_tar <- terra::project(rR0_pip_245_mid_spatial, gorris_tar)
mid_risk_quin <- terra::project(rR0_pip_245_mid_spatial, gorris_quin)

## rasterize
mid_risk_pip_raster <- terra::rasterize(mid_risk_pip, gorris_pip, field = "peakmonths_avgrR0", touches = FALSE)
mid_risk_tar_raster <- terra::rasterize(mid_risk_tar, gorris_tar, field = "peakmonths_avgrR0", touches = FALSE)
mid_risk_quin_raster <- terra::rasterize(mid_risk_quin, gorris_quin, field = "peakmonths_avgrR0", touches = FALSE)

## multiple to create scaled risk
mid_scaled_risk_pip <- gorris_pip *mid_risk_pip_raster
mid_scaled_risk_tar <- gorris_tar *mid_risk_tar_raster
mid_scaled_risk_quin <- gorris_quin *mid_risk_quin_raster

## correct CRS
mid_scaled_risk_pip_correct <- terra::project(mid_scaled_risk_pip, "EPSG:4326")
mid_scaled_risk_tar_correct <- terra::project(mid_scaled_risk_tar, "EPSG:4326")
mid_scaled_risk_quin_correct <- terra::project(mid_scaled_risk_quin, "EPSG:4326")

# not sure if i needed to do that
plot(mid_scaled_risk_quin_correct)
plot(mid_scaled_risk_quin)

### make dataframes
mid_scaled_risk_pip_df <- as.data.frame(mid_scaled_risk_pip_correct, xy = TRUE, na.rm = TRUE)
names(mid_scaled_risk_pip_df)[3] <- "scaled_risk"

mid_scaled_risk_tar_df <- as.data.frame(mid_scaled_risk_tar_correct, xy = TRUE, na.rm = TRUE)
names(mid_scaled_risk_tar_df)[3] <- "scaled_risk"

mid_scaled_risk_quin_df <- as.data.frame(mid_scaled_risk_quin_correct, xy = TRUE, na.rm = TRUE)
names(mid_scaled_risk_quin_df)[3] <- "scaled_risk"

## plot them!
ggplot(mid_scaled_risk_pip_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_pip_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 

ggplot(mid_scaled_risk_tar_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_tar_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 


ggplot(mid_scaled_risk_quin_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_quin_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 

## save them
write.csv(mid_scaled_risk_pip_df, file = "data/processed/temp/map_scaledrisk_245_mid_pip.csv")
write.csv(mid_scaled_risk_tar_df, file = "data/processed/temp/map_scaledrisk_245_mid_tar.csv")
write.csv(mid_scaled_risk_quin_df, file = "data/processed/temp/map_scaledrisk_245_mid_quin.csv")
```


Late century
```{r}
## project
late_risk_pip <- terra::project(rR0_pip_245_late_spatial, gorris_pip)
late_risk_tar <- terra::project(rR0_pip_245_late_spatial, gorris_tar)
late_risk_quin <- terra::project(rR0_pip_245_late_spatial, gorris_quin)

## rasterize
late_risk_pip_raster <- terra::rasterize(late_risk_pip, gorris_pip, field = "peakmonths_avgrR0", touches = FALSE)
late_risk_tar_raster <- terra::rasterize(late_risk_tar, gorris_tar, field = "peakmonths_avgrR0", touches = FALSE)
late_risk_quin_raster <- terra::rasterize(late_risk_quin, gorris_quin, field = "peakmonths_avgrR0", touches = FALSE)

## multiple to create scaled risk
late_scaled_risk_pip <- gorris_pip *late_risk_pip_raster
late_scaled_risk_tar <- gorris_tar *late_risk_tar_raster
late_scaled_risk_quin <- gorris_quin *late_risk_quin_raster

## correct CRS
late_scaled_risk_pip_correct <- terra::project(late_scaled_risk_pip, "EPSG:4326")
late_scaled_risk_tar_correct <- terra::project(late_scaled_risk_tar, "EPSG:4326")
late_scaled_risk_quin_correct <- terra::project(late_scaled_risk_quin, "EPSG:4326")

# not sure if i needed to do that
plot(late_scaled_risk_quin_correct)
plot(late_scaled_risk_quin)

### make dataframes
late_scaled_risk_pip_df <- as.data.frame(late_scaled_risk_pip_correct, xy = TRUE, na.rm = TRUE)
names(late_scaled_risk_pip_df)[3] <- "scaled_risk"

late_scaled_risk_tar_df <- as.data.frame(late_scaled_risk_tar_correct, xy = TRUE, na.rm = TRUE)
names(late_scaled_risk_tar_df)[3] <- "scaled_risk"

late_scaled_risk_quin_df <- as.data.frame(late_scaled_risk_quin_correct, xy = TRUE, na.rm = TRUE)
names(late_scaled_risk_quin_df)[3] <- "scaled_risk"

## plot them!
ggplot(late_scaled_risk_pip_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_pip_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 

ggplot(late_scaled_risk_tar_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_tar_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 


ggplot(late_scaled_risk_quin_df) +
  geom_raster(aes(x = x, y = y, fill = scaled_risk)) +
  scale_fill_viridis_c(option = "D", direction = -1, na.value = "gray95",
    limits = c(0, max(scaled_risk_quin_df$scaled_risk, na.rm = TRUE)),
    oob = scales::squish) +
  coord_sf(xlim = c(-125, -66.5), ylim = c(24, 50), expand = FALSE) +
  theme_void(base_family = "Helvetica") 

## save them
write.csv(late_scaled_risk_pip_df, file = "data/processed/temp/map_scaledrisk_245_late_pip.csv")
write.csv(late_scaled_risk_tar_df, file = "data/processed/temp/map_scaledrisk_245_late_tar.csv")
write.csv(late_scaled_risk_quin_df, file = "data/processed/temp/map_scaledrisk_245_late_quin.csv")
```
